requirements.txt
sqlmodel
fastapi
pytest
allure-pytest
uvicorn
pydantic
aiosqlite
SQLAlchemy
greenlet
httpx
pytest-asyncio
requests
langchain-community
langchain
openai
faiss-cpu
Python-dotenv
tiktoken
langchain-openai